Change to Rating Bar

+++ app/src/main/java/com/example/book_souls_project/api/types/review/ReviewCreateRequest.java
@@ -3,13 +3,13 @@ package com.example.book_souls_project.api.types.review;
 public class ReviewCreateRequest {
     private String bookId;
     private String userId;
-    private int rating;
+    private double rating; // Changed from int to double to support half stars
     private String comment;
 
     // Constructors
     public ReviewCreateRequest() {}
 
-    public ReviewCreateRequest(String bookId, String userId, int rating, String comment) {
+    public ReviewCreateRequest(String bookId, String userId, double rating, String comment) {
         this.bookId = bookId;
         this.userId = userId;
         this.rating = rating;
@@ -23,8 +23,8 @@ public class ReviewCreateRequest {
     public String getUserId() { return userId; }
     public void setUserId(String userId) { this.userId = userId; }
 
-    public int getRating() { return rating; }
-    public void setRating(int rating) { this.rating = rating; }
+    public double getRating() { return rating; }
+    public void setRating(double rating) { this.rating = rating; }
 
     public String getComment() { return comment; }
     public void setComment(String comment) { this.comment = comment; }
